+++ aries-site/src/examples/templates/wizard/TwoColumnWizardExample.js
@@ -248,6 +248,7 @@ export const TwoColumnWizardExample = () => {
         setActiveStep,
         attemptedAdvance,
         setAttemptedAdvance,
+        defaultFormValues,
         error,
         setError,
         formValues,
