+++ aries-site/src/examples/templates/wizard/WizardValidationExample.js
@@ -1,15 +1,9 @@
 import React, { useContext, useEffect, useRef, useState } from 'react';
-import PropTypes from 'prop-types';
 import {
   Box,
-  Button,
   CheckBoxGroup,
-  Footer,
-  Form,
   FormField,
-  Header,
-  Heading,
-  Layer,
+  Grid,
   List,
   RadioButtonGroup,
   ResponsiveContext,
@@ -18,17 +12,17 @@ import {
   TextArea,
   TextInput,
 } from 'grommet';
+import { Checkmark } from 'grommet-icons';
 import {
-  Checkmark,
-  CircleAlert,
-  FormClose,
-  FormNextLink,
-  FormPreviousLink,
-} from 'grommet-icons';
+  CancellationLayer,
+  Error,
+  StepFooter,
+  StepContent,
+  WizardContext,
+  WizardHeader,
+} from './components';
 
-const WizardContext = React.createContext({});
-
-const defaultFormValues = {
+export const defaultFormValues = {
   'text-input-validation': '',
   'radio-button-group-validation': '',
   select: '',
@@ -52,6 +46,8 @@ const validation = [
   {
     validator: values => stepOneValidate(values),
     error: {
+      firstname: '',
+      lastname: '',
       email: '',
       radiobuttongroup: '',
       isValid: true,
@@ -59,51 +55,103 @@ const validation = [
   },
 ];
 
+const stepOneInputs = [
+  (attemptedAdvance, error, formValues, setError) => (
+    <Box>
+      <FormField
+        label="First Name"
+        htmlFor="firstname-validation"
+        name="firstname-validation"
+        error={error.firstname}
+        onChange={() =>
+          attemptedAdvance && setError(stepOneValidate(formValues))
+        }
+      >
+        <TextInput
+          placeholder="Jane"
+          id="firstname-validation"
+          name="firstname-validation"
+        />
+      </FormField>
+      <FormField
+        label="Last Name"
+        htmlFor="lastname-validation"
+        name="lastname-validation"
+        error={error.lastname}
+        onChange={() =>
+          attemptedAdvance && setError(stepOneValidate(formValues))
+        }
+      >
+        <TextInput
+          placeholder="Smith"
+          id="lastname-validation"
+          name="lastname-validation"
+        />
+      </FormField>
+      <FormField
+        label="Email"
+        htmlFor="text-input-validation"
+        name="text-input-validation"
+        error={error.email}
+        onChange={() =>
+          attemptedAdvance && setError(stepOneValidate(formValues))
+        }
+      >
+        <TextInput
+          placeholder="jane.smith@hpe.com"
+          id="text-input-validation"
+          name="text-input-validation"
+          type="email"
+        />
+      </FormField>
+    </Box>
+  ),
+  (attemptedAdvance, error) => (
+    <>
+      <FormField
+        htmlFor="radio-button-group-validation"
+        label="RadioButtonGroup"
+        name="radio-button-group-validation"
+      >
+        <RadioButtonGroup
+          id="radio-button-group-validation"
+          name="radio-button-group-validation"
+          options={['Radio button 1', 'Radio button 2']}
+        />
+      </FormField>
+      {!error.isValid && (
+        <Error>There is an error with one or more inputs.</Error>
+      )}
+    </>
+  ),
+];
+
 const StepOne = () => {
   const { attemptedAdvance, formValues, error, setError } = useContext(
     WizardContext,
   );
-  console.log(formValues);
+  const size = useContext(ResponsiveContext);
+  console.log(error.isValid);
   return (
     <>
       <Box margin={{ bottom: 'medium' }}>
-        <FormField
-          label="Email"
-          htmlFor="text-input-validation"
-          name="text-input-validation"
-          error={error.email}
-          onChange={() =>
-            attemptedAdvance && setError(stepOneValidate(formValues))
-          }
+        <Grid
+          columns={size !== 'small' ? { count: 2, size: 'auto' } : '100%'}
+          rows={[['auto', 'full']]}
+          gap={size !== 'small' ? 'large' : undefined}
+          fill
         >
-          <TextInput
-            placeholder="jane.smith@hpe.com"
-            id="text-input-validation"
-            name="text-input-validation"
-            type="email"
-          />
-        </FormField>
-        <FormField
-          htmlFor="radio-button-group-validation"
-          label="RadioButtonGroup"
-          name="radio-button-group-validation"
-        >
-          <RadioButtonGroup
-            id="radio-button-group-validation"
-            name="radio-button-group-validation"
-            options={['Radio button 1', 'Radio button 2']}
-          />
-        </FormField>
-        {!error.isValid && (
-          <Error>There is an error with one or more inputs.</Error>
-        )}
+          {stepOneInputs.map(input =>
+            input(attemptedAdvance, error, formValues, setError),
+          )}
+        </Grid>
       </Box>
     </>
   );
 };
 
 const StepTwo = () => (
-  <Box margin={{ bottom: 'medium' }}>
+  <Box margin={{ bottom: 'medium' }} width="medium">
     <FormField label="Select" htmlFor="select" name="select">
       <Select
         placeholder="Select item"
@@ -144,7 +192,7 @@ const data = [
 
 const StepThree = () => {
   return (
-    <Box gap="small" margin={{ bottom: 'medium' }}>
+    <Box gap="small">
       <List data={data} pad={{ horizontal: 'none', vertical: 'small' }}>
         {(datum, index) => (
           <Box key={index} direction="row" gap="small" align="center">
@@ -155,14 +203,14 @@ const StepThree = () => {
           </Box>
         )}
       </List>
-      <Text>
+      <Text color="text-strong">
         Include guidance to what will occur when “Finish Wizard" is clicked.
       </Text>
     </Box>
   );
 };
 
-const steps = [
+export const steps = [
   {
     description: `Step one description. Keep each step simple and in chunks 
     easy enough to fit on a single page.`,
@@ -170,14 +218,15 @@ const steps = [
     title: 'Step 1 Title',
   },
   {
-    description: 'Step 2 description.',
+    description: `Step 2 description. Even though this step is single-column, 
+    the width of the footer should be consistent across all steps.`,
     inputs: <StepTwo />,
     title: 'Step 2 Title',
   },
   {
-    description: 'Provide a summary of what was accomplished or configured. ',
+    description: 'Review your configuration details.',
     inputs: <StepThree />,
-    title: 'Finish title',
+    title: 'Review & Create',
   },
 ];
 
@@ -193,7 +242,6 @@ export const WizardValidationExample = () => {
 
   // controls state of cancel layer
   const [open, setOpen] = useState(false);
-  const size = useContext(ResponsiveContext);
 
   // controls error message for active step
   const [error, setError] = useState(
@@ -213,10 +261,11 @@ export const WizardValidationExample = () => {
     setAttemptedAdvance(false);
   }, [activeIndex]);
 
+  const id = 'simple-wizard';
   // scroll to top of step when step changes
   React.useEffect(() => {
     const container = wizardRef.current;
-    const header = document.querySelector('#sticky-header-single-column');
+    const header = document.querySelector(`#${id}`);
     container.scrollTop = -header.getBoundingClientRect().bottom;
   }, [activeIndex, open]);
 
@@ -224,250 +273,28 @@ export const WizardValidationExample = () => {
     <WizardContext.Provider
       value={{
         activeIndex,
+        id,
+        defaultFormValues,
         setActiveIndex,
         activeStep,
         setActiveStep,
         attemptedAdvance,
         setAttemptedAdvance,
         error,
+        ref: wizardRef,
         setError,
+        steps,
         formValues,
         setFormValues,
+        validation,
       }}
     >
-      <Box width={{ max: 'xxlarge' }} margin="auto" fill>
+      <Box fill>
         <WizardHeader setOpen={setOpen} />
-        <Box
-          align="center"
-          pad={size !== 'small' ? { vertical: 'large' } : 'medium'}
-          overflow="auto"
-          ref={wizardRef}
-          flex={size === 'small' ? true : undefined}
-        >
-          <Box width="medium" gap="medium">
-            <StepHeader />
-            <Form
-              // needed to associate form submit button with form
-              // since submit button lives outside form tag
-              id="validation-form"
-              value={formValues}
-              onChange={nextValue => setFormValues(nextValue)}
-              onSubmit={({ value }) => console.log(value)}
-            >
-              {steps[activeIndex].inputs}
-            </Form>
-          </Box>
-        </Box>
+        <StepContent />
         <StepFooter />
       </Box>
       {open && <CancellationLayer onSetOpen={setOpen} />}
     </WizardContext.Provider>
   );
-};
-
-const WizardHeader = ({ setOpen }) => {
-  const size = useContext(ResponsiveContext);
-  const { activeIndex, activeStep, setActiveIndex } = useContext(WizardContext);
-  return (
-    <Header
-      border={{ side: 'bottom', color: 'border-weak' }}
-      pad="small"
-      fill="horizontal"
-      justify="center"
-      responsive={false}
-    >
-      <Box direction="row" flex>
-        {activeStep > 1 && (
-          <Button
-            label={
-              size !== 'small'
-                ? (steps[activeIndex - 1] && steps[activeIndex - 1].title) ||
-                  `Step ${activeStep - 1} Title`
-                : undefined
-            }
-            icon={<FormPreviousLink />}
-            onClick={() => setActiveIndex(activeIndex - 1)}
-          />
-        )}
-      </Box>
-      <Box>
-        <Text color="text-strong" weight="bold">
-          Wizard Title
-        </Text>
-      </Box>
-      <Box direction="row" flex justify="end">
-        <Button
-          label={size !== 'small' ? 'Cancel' : undefined}
-          icon={<FormClose />}
-          reverse
-          onClick={() => setOpen(true)}
-        />
-      </Box>
-    </Header>
-  );
-};
-
-WizardHeader.propTypes = {
-  setOpen: PropTypes.func.isRequired,
-};
-
-const StepHeader = () => {
-  const { activeIndex, activeStep } = useContext(WizardContext);
-  return (
-    <Box id="sticky-header-single-column" gap="medium" flex={false}>
-      <Box>
-        <Text>
-          Step {activeStep} of {steps.length}
-        </Text>
-        <Heading color="text-strong" margin="none">
-          {steps[activeIndex].title || `Step ${activeStep} Title`}
-        </Heading>
-      </Box>
-      <Text size="large">{steps[activeIndex].description}</Text>
-    </Box>
-  );
-};
-
-const StepFooter = () => {
-  const size = useContext(ResponsiveContext);
-  const {
-    activeIndex,
-    setActiveIndex,
-    formValues,
-    setError,
-    setAttemptedAdvance,
-  } = useContext(WizardContext);
-
-  const buttonProps = {
-    fill: size === 'small' ? 'horizontal' : undefined,
-    icon: <FormNextLink />,
-    primary: true,
-    reverse: true,
-  };
-
-  return (
-    <Footer
-      border={{ side: 'top', color: 'border' }}
-      justify="end"
-      pad={
-        size !== 'small'
-          ? { vertical: 'medium' }
-          : { vertical: 'small', horizontal: 'medium' }
-      }
-      width="medium"
-      alignSelf="center"
-    >
-      {activeIndex < steps.length - 1 && (
-        <Button
-          {...buttonProps}
-          label="Next"
-          onClick={() => {
-            // mark that the user is trying to advance, so that onChange
-            // validation will run on any errors in the future
-            setAttemptedAdvance(true);
-
-            let nextIndex = activeIndex + 1;
-            nextIndex = nextIndex <= steps.length - 1 ? nextIndex : activeIndex;
-
-            if (validation[activeIndex]) {
-              // check for errors
-              const validationRes =
-                validation[activeIndex].validator &&
-                validation[activeIndex].validator(formValues);
-              // advance to next step if successful
-              if (validationRes && validationRes.isValid)
-                setActiveIndex(nextIndex);
-              // otherwise, display error and wizard will not advance to
-              // next step
-              else {
-                setError(validationRes);
-              }
-            } else {
-              setActiveIndex(nextIndex);
-            }
-          }}
-        />
-      )}
-      {activeIndex === steps.length - 1 && (
-        <Button
-          {...buttonProps}
-          label="Finish Wizard"
-          form="validation-form-two-column"
-          type="submit"
-        />
-      )}
-    </Footer>
-  );
-};
-
-const CancellationLayer = ({ onSetOpen }) => {
-  const { setFormValues } = useContext(WizardContext);
-  return (
-    <Layer
-      position="center"
-      onClickOutside={() => onSetOpen(false)}
-      onEsc={() => onSetOpen(false)}
-    >
-      <Box pad="large" gap="medium" width="large">
-        <>
-          <Heading color="text-strong" margin="none">
-            Cancel
-          </Heading>
-          <Text color="text-strong">Wizard Title</Text>
-        </>
-        <Text>
-          Cancelling setup will lose all of your progress. Are you sure you want
-          to exit the setup?
-        </Text>
-        <Box
-          as="footer"
-          gap="small"
-          direction="row"
-          align="center"
-          justify="end"
-        >
-          <Button
-            label="No, Continue Wizarding"
-            onClick={() => onSetOpen(false)}
-            secondary
-          />
-          <Button
-            label="Yes, Cancel Wizarding"
-            onClick={() => {
-              onSetOpen(false);
-              setFormValues(defaultFormValues);
-            }}
-            primary
-          />
-        </Box>
-      </Box>
-    </Layer>
-  );
-};
-
-CancellationLayer.propTypes = {
-  onSetOpen: PropTypes.func.isRequired,
-};
-
-const Error = ({ children, ...rest }) => {
-  return (
-    <Box
-      animation="fadeIn"
-      background="validation-critical"
-      margin={{ top: 'small' }}
-      pad="small"
-      round="4px"
-    >
-      <Box direction="row" gap="xsmall" {...rest}>
-        <Box flex={false} margin={{ top: 'hair' }} pad={{ top: 'xxsmall' }}>
-          <CircleAlert size="small" />
-        </Box>
-        <Text size="xsmall">{children}</Text>
-      </Box>
-    </Box>
-  );
-};
-
-Error.propTypes = {
-  children: PropTypes.string,
 };
