+++ aries-site/src/examples/templates/wizard/components/Error.js
@@ -11,9 +11,8 @@ export const Error = ({ children, ...rest }) => {
       margin={{ top: 'small' }}
       pad="small"
       round="4px"
-      {...rest}
     >
-      <Box direction="row" gap="xsmall">
+      <Box direction="row" gap="xsmall" {...rest}>
         <Box flex={false} margin={{ top: 'hair' }} pad={{ top: 'xxsmall' }}>
           <CircleAlert size="small" />
         </Box>
