+++ aries-site/src/examples/templates/wizard/index.js
@@ -1,3 +1,2 @@
 export * from './TwoColumnWizardExample';
 export * from './WizardValidationExample';
-export * from './WizardExample';
