+++ aries-site/src/pages/templates/wizard.mdx
@@ -1,15 +1,12 @@
 import { Example } from '../../layouts';
 import {
-  SimpleWizardExample,
-  StickyHeaderFooterExample,
   TwoColumnWizardExample,
   WizardValidationExample,
-  WizardExample,
 } from '../../examples';
 
 <Example
   code={[
-    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/wizard/WizardExample.js',
+    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/wizard/WizardValidationExample.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/wizard/components/WizardHeader.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/wizard/components/StepHeader.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/wizard/components/StepContent.js',
@@ -22,7 +19,7 @@ import {
   screenContainer
   width="100%"
 >
-  <WizardExample />
+  <WizardValidationExample />
 </Example>
 
 ## Guidance
